refactor(form): make the components and export consistent
