+++ src/components/organisms/Form/FormButton/FormButton.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
-import Button, { IButtonProps } from '../../atoms/Button/Button';
+import Button, { IButtonProps } from '../../../atoms/Button/Button';
 
-import useFormContext from './useFormContext';
+import useFormContext from '../useFormContext';
 
 const FormButton: FC<IButtonProps> = props => {
   const { invalid, handleSubmit } = useFormContext();
